requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib
bcrypt==4.0.1
pandas
numpy
scikit-learn
joblib
gdown
python-dotenv
python-multipart
gunicorn
uvicorn